pyproject.toml
[project]
name = "dao-ai-ui"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dao-ai>=0.0.28",
    "databricks-sdk>=0.73.0",
    "mlflow>=3.6.0",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
]
